// generated/openapi_geolocation.ts
export interface paths {
    "/v1/geolocation/mission/{id}/raw": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Query mission
         * @description Returns a single mission specified by its unique ID
         */
        get: operations["getV1GeolocationMissionByIdRaw"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v1/geolocation/missions/raw": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Query missions
         * @description Returns a list of missions specified by their unique IDs
         */
        get: operations["getV1GeolocationMissionsRaw"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v1/geolocation/missions/tda": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post: operations["postV1GeolocationMissionsTda"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v1/auth": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["getV1Auth"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v1/atest": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["getV1Atest"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: never;
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
    getV1GeolocationMissionByIdRaw: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                id: number;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        db: {
                            id: number;
                            source: string;
                            timestamp: Record<string, never>;
                            missionid: number;
                            lng: number;
                            lat: number;
                            has_chauffeur_reached_end: boolean;
                        } | null;
                        waynium: unknown;
                        elastic: unknown;
                    };
                    "multipart/form-data": {
                        db: {
                            id: number;
                            source: string;
                            timestamp: Record<string, never>;
                            missionid: number;
                            lng: number;
                            lat: number;
                            has_chauffeur_reached_end: boolean;
                        } | null;
                        waynium: unknown;
                        elastic: unknown;
                    };
                    "text/plain": {
                        db: {
                            id: number;
                            source: string;
                            timestamp: Record<string, never>;
                            missionid: number;
                            lng: number;
                            lat: number;
                            has_chauffeur_reached_end: boolean;
                        } | null;
                        waynium: unknown;
                        elastic: unknown;
                    };
                };
            };
        };
    };
// ...
}